// src/interfaces/IGlobalTheme.ts
import { THEME } from './ETheme';

export interface IGlobalTheme {
  theme: THEME;
  primaryColor: string;
  secondaryColor: string;
  subSecondaryColor: string;
  markdownParagraphColor: string;
  primaryTextColor: string;
  secondaryTextColor: string;
  subSecondaryTextColor: string;
  subTabColor: string;
  subTabHoverColor: string;
  primaryButtonColor: string;
  secondaryButtonColor: string;
  primaryIconColor: string;
  secondaryIconColor: string;
  selectedFileColor: string;
  hoverFileColor: string;
  scrollbarColor: string;
  green: string;
  blue: string;
  white: string;
  link: string;
  toolBarColor: string;
}